Extract common graphql request handling into abstract parent.
